refactor: removed all cmds of default behavior and the global vars used by those
The previously defined additional commands and global vars that were
added in the plugin file didn't serve their purpose. Once defined in
there the commands' behavior can't really be changed by changing the
variables through the cmdline or in the config file.

It is better for the user to define these commands and variables in the
config file on their own based on their needs. How? -> Instructions are
in the README.